major refactor - broke key functions out into their separate components
